+++ app/src/main/java/tiniakovdev/com/API.kt
@@ -1,0 +1,5 @@
+package tiniakovdev.com
+
+object API {
+    const val KEY = "cecf3b811b2cab303b0e95b085b03608"
+}
