+++ app/src/main/java/tiniakovdev/com/data/MainRepository.kt
@@ -1,22 +1,11 @@
-package tiniakovdev.com
+package tiniakovdev.com.data
 
+import tiniakovdev.com.R
+import tiniakovdev.com.domain.Film
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.LinearLayoutManager
-import tiniakovdev.com.databinding.FragmentHomeBinding
-import java.util.*
+class MainRepository {
 
-
-class HomeFragment : Fragment() {
-
-    private lateinit var filmAdapter: FilmListRecyclerAdapter
-    private lateinit var binding: FragmentHomeBinding
-
-    private val filmDataBase = listOf(
+    val filmsDataBase = listOf(
         Film(
             "Гладиатор", R.drawable.gladiator, "В великой Римской империи не было" +
                     " военачальника, равного генералу Максимусу. Непобедимые легионы, которыми" +
@@ -117,69 +106,4 @@ class HomeFragment : Fragment() {
         )
     )
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        AnimationHelper.performFragmentCircularRevealAnimation(
-            binding.homeFragmentRoot,
-            requireActivity(),
-            1
-        )
-
-        initSearchView()
-        initRecycler()
-        filmAdapter.addItems(filmDataBase)
-    }
-
-    private fun initSearchView() {
-        binding.searchView.setOnClickListener {
-            binding.searchView.isIconified = false
-        }
-
-        binding.searchView.setOnQueryTextListener(object :
-            androidx.appcompat.widget.SearchView.OnQueryTextListener {
-            override fun onQueryTextSubmit(query: String?): Boolean {
-                return true
-            }
-
-            override fun onQueryTextChange(newText: String): Boolean {
-                if (newText.isEmpty()) {
-                    filmAdapter.addItems(filmDataBase)
-                    return true
-                }
-                val result = filmDataBase.filter {
-                    it.title.lowercase(Locale.getDefault())
-                        .contains(newText.lowercase(Locale.getDefault()))
-                }
-                filmAdapter.addItems(result)
-                return true
-            }
-        })
-        initRecycler()
-        filmAdapter.addItems(filmDataBase)
-    }
-
-    private fun initRecycler() {
-        binding.mainRecycler.apply {
-            filmAdapter =
-                FilmListRecyclerAdapter(object : FilmListRecyclerAdapter.OnItemClickListener {
-                    override fun click(film: Film) {
-                        (requireActivity() as MainActivity).launchDetailsFragment(film)
-                    }
-                })
-            adapter = filmAdapter
-            layoutManager = LinearLayoutManager(requireContext())
-            val decoration = TopSpacingItemDecoration(8)
-            addItemDecoration(decoration)
-        }
-    }
 }
-
-
