+++ app/src/main/java/tiniakovdev/com/domain/Interactor.kt
@@ -1,0 +1,8 @@
+package tiniakovdev.com.domain
+
+import tiniakovdev.com.data.MainRepository
+
+class Interactor(val repository: MainRepository) {
+
+    fun getFilmsDB(): List<Film> = repository.filmsDataBase
+}
