+++ app/src/main/java/tiniakovdev/com/view/fragments/DetailsFragment.kt
@@ -1,4 +1,4 @@
-package tiniakovdev.com
+package tiniakovdev.com.view.fragments
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,7 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import tiniakovdev.com.R
 import tiniakovdev.com.databinding.FragmentDetailsBinding
+import tiniakovdev.com.domain.Film
 
 
 class DetailsFragment : Fragment() {
