+++ app/src/main/java/tiniakovdev/com/view/fragments/WatchLaterFragment.kt
@@ -1,4 +1,4 @@
-package tiniakovdev.com
+package tiniakovdev.com.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import tiniakovdev.com.databinding.FragmentWatchLaterBinding
+import tiniakovdev.com.utils.AnimationHelper
 
 class WatchLaterFragment : Fragment() {
