+++ app/src/main/java/tiniakovdev/com/view/rv_adapter/FilmListRecyclerAdapter.kt
@@ -1,9 +1,12 @@
-package tiniakovdev.com
+package tiniakovdev.com.view.rv_adapter
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import tiniakovdev.com.view.rv_viewholders.FilmViewHolder
+import tiniakovdev.com.R
+import tiniakovdev.com.domain.Film
 
 class FilmListRecyclerAdapter(private val clickListener: OnItemClickListener) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
