+++ app/src/main/res/layout/film_item.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="150dp">
 
-    <tiniakovdev.com.RatingDonutView
+    <tiniakovdev.com.view.customviews.RatingDonutView
         android:id="@+id/rating_donut"
         android:layout_width="40dp"
         android:layout_height="40dp"
