+++ app/src/main/res/layout/fragment_favorite.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/shape"
     android:visibility="invisible"
-    tools:context=".FavoritesFragment"
+    tools:context=".view.fragments.FavoritesFragment"
     tools:visibility="visible">
 
     <androidx.recyclerview.widget.RecyclerView
