+++ app/src/main/res/layout/fragment_watch_later.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/shape"
     android:visibility="invisible"
-    tools:context=".WatchLaterFragment"
+    tools:context=".view.fragments.WatchLaterFragment"
     tools:visibility="visible">
 
 </FrameLayout>
